handle local psuedo-fetcher more robustly
In spacemacs, quelpa recipes can use the spacemacs-specific "local"
pseudofetcher, e.g., a recipe might be
```
(ess :location (recipe
                     :fetcher local
                     :files ("*.el" "lisp/*.el" ; copied from MELPA recipe
                                  "doc/ess.texi"
                                  ("etc" "etc/*")
                                  ("obsolete" "lisp/obsolete/*")
                                  (:exclude "etc/other"))))
```

In practice, spacemacs finds the explicit path (the package should live under
/local/ in the associated layer), and then constructs a fresh recipe on the fly
using backquoting like
```
`(recipe :fetcher file :path ,path)
```

Unfortunately, this clobbers any other fields that were in the recipe which may
be necessary for constructing the package (e.g., the `:files` keyword from the
`ess` example above would effectively be clipped from the recipe.

This commit tries to address this shortcoming. It still finds the explicit path
as before, but rather than constructing a fresh recipe, it manipulates
the *existing* recipe. This way, any manually defined properties will persist.

Finally, when location is defined, regardless of whether we manipulate it or
not, set it in the obj being constructed